move courseDetail to components folder
