+++ src/components/header/header.tsx
@@ -8,20 +8,20 @@ import InputMask from "react-input-mask";
 import { useRouter } from "next/router";
 import { useRecoilState, SetterOrUpdater } from "recoil";
 import { ref, uploadBytesResumable, getDownloadURL } from "firebase/storage";
-import { AuthContext } from "../contexts/authContext";
-import { getUserData, updateAvatar, updateBearMoney, updateTeacherData } from "../utils/firestore";
-import { orderQtyState, bearMoneyState, showMemberModalState } from "../utils/recoil";
-import { storage } from "../../lib/firebase";
-import BearLogo from "../../public/bear-logo2.png";
-import CartLogo from "../../public/cart.png";
-import MemberLogo from "../../public/member.png";
-import MoneyIcon from "../../public/newMoney.png";
-import PlusMoneyIcon from "../../public/add.png";
-import MoneyBear from "../../public/money.png";
-import Loading from "../../public/loading.gif";
-import MenuIcon from "../../public/menu.png";
-import Editor from "./editor/editor";
-import Modal from "./modal";
+import { AuthContext } from "../../contexts/authContext";
+import { getUserData, updateAvatar, updateBearMoney, updateTeacherData } from "../../utils/firestore";
+import { orderQtyState, bearMoneyState, showMemberModalState } from "../../utils/recoil";
+import { storage } from "../../../lib/firebase";
+import BearLogo from "../../../public/bear-logo2.png";
+import CartLogo from "../../../public/cart.png";
+import MemberLogo from "../../../public/member.png";
+import MoneyIcon from "../../../public/newMoney.png";
+import PlusMoneyIcon from "../../../public/add.png";
+import MoneyBear from "../../../public/money.png";
+import Loading from "../../../public/loading.gif";
+import MenuIcon from "../../../public/menu.png";
+import Editor from "../editor/editor";
+import Modal from "../modal";
 
 const Wrapper = styled.header`
   display: flex;
